# backend/swagger.json
{
    "swagger": "3.0",
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "AgileOne CRUD",
        "description": "Recruiting Application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:2000/api/",
            "description": "Dev-Local"
        },
        {
            "url": "http://34.224.218.231:3000/api/",
            "description": "Dev-Server"
        }
    ],
    "components": {        
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "schemes": ["http","https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/admin/get_userdata": {
            "post": {
                "tags": ["Admin"],
                "description": "Get Userdata",
                "operationId": "get_userdata",
                "summary": "Get User Details in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "loginType": "admin/employee",
                                "user_id": "User ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_teamleads": {
            "post": {
                "tags": ["Admin"],
                "description": "Get All Team Leads",
                "operationId": "get_teamleads",
                "summary": "Get All Teamleads in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "user_id": "User ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/get_tlemployees": {
            "post": {
                "tags": ["Admin"],
                "description": "Get All Employees Under Team Leads",
                "operationId": "get_tlemployees",
                "summary": "Get All Employees Under Teamleads in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "user_id": "Team Lead ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/get_dashboarddata": {
            "post": {
                "tags": ["Admin"],
                "description": "Get Dashboard Data",
                "operationId": "get_dashboarddata",
                "summary": "Get Dashboard Data for all roles in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "role": "Role ID from get_roles api",
                                "user_id": "User ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}